DC-607: minor edits renaming files and utilising common function to generate the from address
